+++ src/pages/MyPage/MyPage.module.scss
@@ -39,90 +39,10 @@
   gap: 1.2rem;
 }
 
-.input-profile {
-  display: flex;
-  flex-direction: column;
-  gap: 3.2rem;
+.input-password {
   width: 62rem;
-  height: 35.5rem;
+  height: 45.4rem;
+  flex-shrink: 0;
   border-radius: 0.8rem;
   background: $color-white;
-  padding: 2.4rem;
-
-  & h2 {
-    padding-top: 0.4rem;
-    font-size: 2.4rem;
-    font-weight: 700;
-    @include mobile {
-      font-size: 2rem;
-    }
-  }
-}
-
-.input-profile-inputs {
-  display: flex;
-  flex-direction: row;
-  gap: 1.6rem;
-}
-
-.emailandnickname {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  gap: 2rem;
-}
-
-.input {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-
-  & input {
-    height: 4.8rem;
-    border-radius: 0.6rem;
-    border: 0.1rem solid $color-gray-30;
-    padding: 1.5rem;
-    font-size: 1.6rem;
-    font-weight: 400;
-    @include mobile {
-      font-size: 1.4rem;
-      height: 4.2rem;
-    }
-  }
-}
-
-.pictures {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 18.2rem;
-  height: 18.2rem;
-  background-color: $color-gray-15;
-  border-radius: 0.6rem;
-  flex-shrink: 0;
-  position: relative;
-  &:hover {
-    cursor: pointer;
-    background-color: $color-gray-30;
-  }
-
-  & img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
-  & .remove-button {
-    position: absolute;
-    top: 1rem;
-    right: 1rem;
-    width: 1.5rem;
-    height: 1.5rem;
-  }
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
 }
